+++ src/main/java/com/oitsjustjose/geolosys/Geolosys.java
@@ -41,6 +41,8 @@ import org.apache.logging.log4j.Logger;
 
 import javax.annotation.Nonnull;
 import java.io.*;
+import java.util.ArrayList;
+import java.util.List;
 
 @Mod(modid = Geolosys.MODID, name = "Geolosys", version = Geolosys.VERSION, acceptedMinecraftVersions = "1.12", dependencies = "after:immersiveengineering@[0.12,);")
 public class Geolosys